Create facades in the Archethic module for the web
