+++ lib/archethic_web/components/pagination.ex
@@ -1,4 +1,6 @@
 defmodule ArchethicWeb.Pagination do
+  @moduledoc false
+
   use Phoenix.Component
 
   @doc """
